+++ tests/Swagger/Serializer/DocumentationNormalizerV3Test.php
@@ -111,8 +111,8 @@ class DocumentationNormalizerV3Test extends TestCase
         $resourceMetadataFactoryProphecy->create(Dummy::class)->shouldBeCalled()->willReturn($dummyMetadata);
 
         $propertyMetadataFactoryProphecy = $this->prophesize(PropertyMetadataFactoryInterface::class);
-        $propertyMetadataFactoryProphecy->create(Dummy::class, 'id')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_INT), 'This is an id.', true, false, null, null, null, true));
-        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, []));
+        $propertyMetadataFactoryProphecy->create(Dummy::class, 'id')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_INT), 'This is an id.', true, false, null, null, null, true, null, null, null, null, null, null, null, ['minLength' => 3, 'maxLength' => 20, 'pattern' => '^dummyPattern$']));
+        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, [], null, null, null, null));
         $propertyMetadataFactoryProphecy->create(Dummy::class, 'description')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is an initializable but not writable property.', true, false, true, true, false, false, null, null, [], null, true));
         $propertyMetadataFactoryProphecy->create(Dummy::class, 'dummyDate')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_OBJECT, true, \DateTime::class), 'This is a \DateTimeInterface object.', true, true, true, true, false, false, null, null, []));
 
@@ -375,6 +375,9 @@ class DocumentationNormalizerV3Test extends TestCase
                                 'type' => 'integer',
                                 'description' => 'This is an id.',
                                 'readOnly' => true,
+                                'minLength' => 3,
+                                'maxLength' => 20,
+                                'pattern' => '^dummyPattern$',
                             ]),
                             'name' => new \ArrayObject([
                                 'type' => 'string',
@@ -483,7 +486,7 @@ class DocumentationNormalizerV3Test extends TestCase
             $legacy ? $nameConverter : null,
             true,
             'oauth2',
-            'application',
+            'authorizationCode',
             '/oauth/v2/token',
             '/oauth/v2/auth',
             ['scope param'],
@@ -554,11 +557,14 @@ class DocumentationNormalizerV3Test extends TestCase
                 'securitySchemes' => [
                     'oauth' => [
                         'type' => 'oauth2',
-                        'description' => 'OAuth client_credentials Grant',
-                        'flow' => 'application',
-                        'tokenUrl' => '/oauth/v2/token',
-                        'authorizationUrl' => '/oauth/v2/auth',
-                        'scopes' => ['scope param'],
+                        'description' => 'OAuth 2.0 authorization code Grant',
+                        'flows' => [
+                            'authorizationCode' => [
+                                'tokenUrl' => '/oauth/v2/token',
+                                'authorizationUrl' => '/oauth/v2/auth',
+                                'scopes' => ['scope param'],
+                            ],
+                        ],
                     ],
                 ],
             ],